requirements.txt
asn1crypto==0.24.0
biopython==1.81
certifi==2021.10.8
charset-normalizer==2.0.12
entrypoints==0.3
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
keyring==17.1.1
keyrings.alt==3.1.1
numpy==1.21.5
pycrypto==2.6.1
pyxdg==0.25
requests==2.28.2
SecretStorage==2.3.1
six==1.12.0
SQLAlchemy==1.4.35
typing-extensions==4.1.1
urllib3==1.26.9
zipp==3.7.0
